updating with all new drive functionality
includes drive forward, backward, left, and right. Also includes a 180-degree rotate. There are also 2 helper functions, setMotorSpeed  and dynamic braking. setMotorSpeed is used to control the direction and speed of a specific motor. Dynamic braking is used to slow down the motor by setting the motors to a speed proportionally to the original speed, in the opposite direction.